Add support for aspnetcorerazor filetypes
If you are using Microsoft's [C# plugin](https://marketplace.visualstudio.com/items?itemName=ms-vscode.csharp) it changes `Razor` filetypes to be called `aspnetcorerazor` with some enhanced support. The files are the same, but the type name is different, so I've added that type as well. I've tested locally and it seems to work just as expected.